posbackend/requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dateparser==0.7.2
Django==2.2.13
django-crispy-forms==1.8.0
django-debug-toolbar==2.1
django-discover-runner==1.0
django-model-utils==4.0.0
django-rest-swagger==2.2.0
django-role-permissions==2.2.1
django-taggit==1.3.0
djangorestframework==3.13.1
fake-factory==9999.9.9
Faker==2.0.4
idna==2.8
itypes==1.2.0
jet-bridge-base==0.7.7
jet-django==1.0.2
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
Pillow==8.4.0
prompt-toolkit==2.0.9
psycopg2-binary==2.9.2
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2018.5
requests==2.22.0
simplejson==3.17.6
six==1.13.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
text-unidecode==1.3
tzlocal==2.0.0
uritemplate==4.1.1
urllib3==1.25.7
wcwidth==0.1.7
